+++ app/Main.hs
@@ -1,15 +1,28 @@
 {-# LANGUAGE DeriveDataTypeable #-}
+{-# LANGUAGE LambdaCase #-}
+{-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RecordWildCards #-}
+{-# LANGUAGE ViewPatterns #-}
 module Main where
 
+import           Control.Applicative            ( (<|>) )
+import           Control.Exception.Safe         ( try )
 import           Control.Monad                  ( forM_ )
-import           Data.Foldable                  ( asum )
+import           Data.Aeson                     ( (.:?)
+                                                , FromJSON(..)
+                                                , withObject
+                                                )
 import           Data.Maybe                     ( fromMaybe )
 import           Data.Time                      ( Day
                                                 , defaultTimeLocale
                                                 , formatTime
                                                 )
 import           Data.Version                   ( showVersion )
+import           Data.Yaml                      ( prettyPrintParseException )
+import           Data.Yaml.Config               ( ignoreEnv
+                                                , loadYamlSettings
+                                                )
 import           System.Console.CmdArgs         ( (&=)
                                                 , Data
                                                 , Typeable
@@ -26,7 +39,9 @@ import           System.Console.CmdArgs         ( (&=)
                                                 , summary
                                                 , typ
                                                 )
+import           System.Directory               ( doesFileExist )
 import           System.Environment             ( lookupEnv )
+import           System.Environment.XDG.BaseDir ( getUserConfigFile )
 import           System.Exit                    ( exitFailure )
 import           System.IO                      ( hPutStrLn
                                                 , stderr
@@ -42,19 +57,10 @@ import qualified Data.Text                     as T
 
 main :: IO ()
 main = do
-    Args {..}      <- cmdArgs argSpec
-    journalFileEnv <- lookupEnv "LEDGER_FILE"
-    apiKeyEnv      <- lookupEnv "ALPHAVANTAGE_KEY"
-    apiKey         <- case asum [apiKey_, apiKeyEnv] of
-        Just k -> return k
-        Nothing ->
-            logError
-                    "Error: Pass an AlphaVantage API Key with `-a` or $ALPHAVANTAGE_KEY."
-                >> exitFailure
-
-    let journalFile =
-            fromMaybe "~/.hledger.journal" $ asum [journalFile_, journalFileEnv]
-        cfg = Config $ T.pack apiKey
+    cfgArgs        <- cmdArgs argSpec
+    cfgFile        <- loadConfigFile
+    AppConfig {..} <- mergeArgsEnvCfg cfgFile cfgArgs
+    let cfg = Config $ T.pack apiKey
     (commodities, start, end) <- getCommoditiesAndDateRange
         (T.pack <$> excludedCurrencies)
         journalFile
@@ -62,8 +68,7 @@ main = do
         then do
             prices <- fetchPrices cfg commodities start end rateLimit
             if null prices
-                then logError
-                    "Error: No price directives were able to be fetched."
+                then logError "No price directives were able to be fetched."
                 else LBS.writeFile outputFile $ makePriceDirectives prices
         else do
             putStrLn
@@ -77,41 +82,124 @@ main = do
   where
     showDate :: Day -> String
     showDate = formatTime defaultTimeLocale "%Y-%m-%d"
-    logError :: String -> IO ()
-    logError = hPutStrLn stderr
 
 
-data Args = Args
-    { apiKey_            :: Maybe String
+logError :: String -> IO ()
+logError = hPutStrLn stderr . ("[ERROR] " <>)
+
+
+
+-- CONFIGURATION
+
+data AppConfig = AppConfig
+    { apiKey             :: String
     , rateLimit          :: Bool
-    , journalFile_       :: Maybe FilePath
+    , journalFile        :: FilePath
     , outputFile         :: FilePath
     , excludedCurrencies :: [String]
     , dryRun             :: Bool
+    }
+    deriving (Show, Eq)
+
+defaultExcludedCurrencies :: [String]
+defaultExcludedCurrencies = ["$", "USD"]
+
+-- | Merge the Arguments, Environmental Variables, & Configuration File
+-- into an 'AppConfig.
+--
+-- Arguments override environmental variables, which overrides the
+-- configuration file.
+mergeArgsEnvCfg :: ConfigFile -> Args -> IO AppConfig
+mergeArgsEnvCfg ConfigFile {..} Args {..} = do
+    envJournalFile <- lookupEnv "LEDGER_FILE"
+    envApiKey      <- lookupEnv "ALPHAVANTAGE_KEY"
+    apiKey         <- case argApiKey <|> envApiKey <|> cfgApiKey of
+        Just k -> return k
+        Nothing ->
+            logError
+                    "Pass an AlphaVantage API Key with `-a` or $ALPHAVANTAGE_KEY."
+                >> exitFailure
+    let journalFile =
+            fromMaybe "~/.hledger.journal" $ argJournalFile <|> envJournalFile
+        rateLimit =
+            fromMaybe True $ either (const cfgRateLimit) Just argRateLimit
+        excludedCurrencies =
+            if argExcludedCurrencies == defaultExcludedCurrencies
+                then fromMaybe defaultExcludedCurrencies cfgExcludedCurrencies
+                else argExcludedCurrencies
+        outputFile = argOutputFile
+        dryRun     = argDryRun
+    return AppConfig { .. }
+
+
+data ConfigFile = ConfigFile
+    { cfgApiKey             :: Maybe String
+    , cfgRateLimit          :: Maybe Bool
+    , cfgExcludedCurrencies :: Maybe [String]
+    }
+    deriving (Show, Eq)
+
+instance FromJSON ConfigFile where
+    parseJSON = withObject "ConfigFile" $ \o -> do
+        cfgApiKey             <- o .:? "api-key"
+        cfgRateLimit          <- o .:? "rate-limit"
+        cfgExcludedCurrencies <- o .:? "exclude"
+        return ConfigFile { .. }
+
+loadConfigFile :: IO ConfigFile
+loadConfigFile = do
+    configFile <- getUserConfigFile "hledger-stockquotes" "config.yaml"
+    hasConfig  <- doesFileExist configFile
+    if hasConfig
+        then try (loadYamlSettings [configFile] [] ignoreEnv) >>= \case
+            Left (lines . prettyPrintParseException -> errorMsg) ->
+                hPutStrLn stderr "[WARN] Invalid Configuration File Format:"
+                    >> mapM_ (hPutStrLn stderr . ("\t" <>)) errorMsg
+                    >> return defaultConfig
+            Right c -> return c
+        else return defaultConfig
+  where
+    defaultConfig :: ConfigFile
+    defaultConfig = ConfigFile Nothing Nothing Nothing
+
+
+
+data Args = Args
+    { argApiKey             :: Maybe String
+    , argRateLimit          :: Either () Bool
+    , argJournalFile        :: Maybe FilePath
+    , argOutputFile         :: FilePath
+    , argExcludedCurrencies :: [String]
+    , argDryRun             :: Bool
     }
     deriving (Data, Typeable, Show, Eq)
 
 argSpec :: Args
 argSpec =
     Args
-            { apiKey_            =
+            { argApiKey             =
                 Nothing
                 &= help "Your AlphaVantage API key. Default: $ALPHAVANTAGE_KEY"
                 &= explicit
                 &= name "api-key"
                 &= name "a"
                 &= typ "ALPHAVANTAGE_KEY"
-            , rateLimit          = enum
-                                       [ True
-                                       &= help "Apply rate-limting for the API"
-                                       &= ignore
-                                       , False
-                                       &= help "Disable rate-limiting for the API"
-                                       &= explicit
-                                       &= name "no-rate-limit"
-                                       &= name "n"
-                                       ]
-            , journalFile_       =
+            , argRateLimit          = enum
+                [ Left ()
+                &= help "Fall back to the configuration file, or True."
+                &= ignore
+                , Right True
+                &= help "Apply rate-limting for the API"
+                &= explicit
+                &= name "rate-limit"
+                &= name "r"
+                , Right False
+                &= help "Disable rate-limiting for the API"
+                &= explicit
+                &= name "no-rate-limit"
+                &= name "n"
+                ]
+            , argJournalFile        =
                 Nothing
                 &= help
                        "Journal file to read commodities from. Default: $LEDGER_FILE or ~/.hledger.journal"
@@ -119,7 +207,7 @@ argSpec =
                 &= name "journal-file"
                 &= name "f"
                 &= typ "FILE"
-            , outputFile         =
+            , argOutputFile         =
                 "prices.journal"
                 &= help
                        "File to write prices into. Existing files will be overwritten. Default: prices.journal"
@@ -127,10 +215,11 @@ argSpec =
                 &= name "output-file"
                 &= name "o"
                 &= typ "FILE"
-            , excludedCurrencies = ["$", "USD"] &= args &= typ
-                                       "EXCLUDED_CURRENCY ..."
-            , dryRun = False &= explicit &= name "dry-run" &= name "d" &= help
-                "Print the commodities and dates that would be processed."
+            , argExcludedCurrencies = defaultExcludedCurrencies &= args &= typ
+                                          "EXCLUDED_CURRENCY ..."
+            , argDryRun             =
+                False &= explicit &= name "dry-run" &= name "d" &= help
+                    "Print the commodities and dates that would be processed."
             }
         &= summary
                (  "hledger-stockquotes v"
@@ -151,7 +240,7 @@ argSpec =
                , "By default, we find all non-USD commodities in your "
                , "journal file and query AlphaVantage for their stock prices "
                , "over the date range used in the journal file. Currently, we "
-               , "only support public U.S. equities & do not call out to AlphVantage's"
+               , "only support public U.S. equities & do not call out to AlphaVantage's"
                , "FOREX or Crypto API routes. If you have commodities that are "
                , "not supported by AlphaVantage, hledger-stockquotes will output "
                , "an error when attempting to processing them. To avoid processing "
@@ -163,7 +252,7 @@ argSpec =
                , ""
                , "API LIMITS"
                , ""
-               , "AlphVantage's API limits users to 5 requests per minute. We respect "
+               , "AlphaVantage's API limits users to 5 requests per minute. We respect "
                , "this limit by waiting for 60 seconds after every 5 commities we process. "
                , "You can ignore the rate-limiting by using the `-n` flag, but "
                , "requests are more likely to fail. You can use the `-d` flag to print "
@@ -190,6 +279,22 @@ argSpec =
                , ""
                , "Instead of passing the `-a` flag with your AlphaVantage API key, "
                , "you can set the ALPHAVANTAGE_KEY environmental variable instead."
+               , ""
+               , ""
+               , "CONFIGURATION FILE"
+               , ""
+               , "If you have common options you constantly pass to the application, "
+               , "you can specify them in a YAML configuration file. We attempt "
+               , "to parse a configuration file in $XDG_CONFIG_HOME/hledger-stockquotes/config.yaml. "
+               , "It currently supports the following top-level keys: "
+               , ""
+               , "- `api-key`:      (string) Your AlphaVantage API Key"
+               , "- `exclude`:      (list of strings) Currencies to Exclude"
+               , "- `rate-limit`:   (bool) Obey AlphaVantage's Rate Limit"
+               , ""
+               , "Environmental variables will overide any config file options, "
+               , "and CLI flags will override both environmental variables & "
+               , "config file options."
                , ""
                , ""
                , "USAGE EXAMPLES"